requirements.txt
requests
beautifulsoup4
python-frontmatter
markdown2
Jinja2
openai 